backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery==5.4.0
redis==5.2.1
python-multipart==0.0.20
websockets==14.1
onnxruntime==1.21.0
scenedetect[opencv]==0.6.4
Pillow==11.1.0
openai==1.59.5
numpy==2.2.1
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
faster-whisper==1.1.1
requests==2.32.5
packaging==25.0
